+++ backend/src/typeorm/connection.ts
@@ -50,7 +50,6 @@ export class Connection {
     } catch (error) {
       // eslint-disable-next-line no-console
       console.log(error)
-      console.log('catch from createConnection')
       return null
     }
   }
